Restructuring manage device tab
The current layout of the "Manage device" tab
in "Settings" is suboptimal, and should be
improved, also to allow for a correct placement
of the "Change device password" button.

Since items in the translation json (locales/en)
should either be added or removed (not modified),
a fresh "deviceSettings.title" was added.

Also, since both device firmware and bluetooth
firmware will be located in the same section,
a "bluetoothFirmware" section has been added in
order to be able to separate (device) firmware
from bluetooth firmware. Note that this only
affects the "title" and "firmwareVersion" field.