frontend: update add-account layout
Drop the white box around the add-account workflow to make it more
consistent with the rest of the app.

The white box is still used in a few different components.

Other changes:

Removed some custom styling and rewrote using standard
components such as ViewHeader and ViewButtons. This has the nice
side effect that the layout is more consistent with other views.

Changed to nicer looking animated checkmark icon in success step.

Fixed minor layout jump between the steps in the workflow. Reason
was the dropdown and input field had slightly different height and
margin.